Rename DTOs folder rename "DTOs" to "CollaborationModels"
