+++ NuKeeper.Abstractions/CollaborationModels/UserPermissions.cs
@@ -1,4 +1,4 @@
-namespace NuKeeper.Abstractions.DTOs
+namespace NuKeeper.Abstractions.CollaborationModels
 {
     public class UserPermissions
     {
