+++ NuKeeper.Abstractions/CollaborationPlatform/ICollaborationPlatform.cs
@@ -1,7 +1,7 @@
 using NuKeeper.Abstractions.Configuration;
-using NuKeeper.Abstractions.DTOs;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using NuKeeper.Abstractions.CollaborationModels;
 
 namespace NuKeeper.Abstractions.CollaborationPlatform
 {
