+++ NuKeeper.Abstractions/Configuration/RepositorySettings.cs
@@ -1,5 +1,5 @@
 using System;
-using NuKeeper.Abstractions.DTOs;
+using NuKeeper.Abstractions.CollaborationModels;
 
 namespace NuKeeper.Abstractions.Configuration
 {
