+++ NuKeeper.BitBucket/BitbucketForkFinder.cs
@@ -1,8 +1,8 @@
 using System;
 using System.Threading.Tasks;
+using NuKeeper.Abstractions.CollaborationModels;
 using NuKeeper.Abstractions.CollaborationPlatform;
 using NuKeeper.Abstractions.Configuration;
-using NuKeeper.Abstractions.DTOs;
 using NuKeeper.Abstractions.Logging;
 
 namespace NuKeeper.BitBucket
