+++ NuKeeper.Tests/Engine/PackageUpdateSelectionTests.cs
@@ -4,8 +4,8 @@ using System.Linq;
 using System.Threading.Tasks;
 using NSubstitute;
 using NuKeeper.Abstractions;
+using NuKeeper.Abstractions.CollaborationModels;
 using NuKeeper.Abstractions.Configuration;
-using NuKeeper.Abstractions.DTOs;
 using NuKeeper.Abstractions.Logging;
 using NuKeeper.Engine.Packages;
 using NuKeeper.Inspection.Sort;
