+++ NuKeeper/Engine/CollaborationEngine.cs
@@ -1,12 +1,12 @@
 using LibGit2Sharp;
 using NuKeeper.Abstractions.CollaborationPlatform;
 using NuKeeper.Abstractions.Configuration;
-using NuKeeper.Abstractions.DTOs;
 using NuKeeper.Abstractions.Formats;
 using NuKeeper.Abstractions.Logging;
 using NuKeeper.Inspection.Files;
 using System;
 using System.Threading.Tasks;
+using NuKeeper.Abstractions.CollaborationModels;
 
 namespace NuKeeper.Engine
 {
