+++ NuKeeper/Engine/GitRepositoryEngine.cs
@@ -2,9 +2,9 @@ using System;
 using System.IO;
 using System.Threading.Tasks;
 using LibGit2Sharp;
+using NuKeeper.Abstractions.CollaborationModels;
 using NuKeeper.Abstractions.CollaborationPlatform;
 using NuKeeper.Abstractions.Configuration;
-using NuKeeper.Abstractions.DTOs;
 using NuKeeper.Abstractions.Inspections.Files;
 using NuKeeper.Abstractions.Logging;
 using NuKeeper.Git;
