+++ NuKeeper/Engine/Packages/ExistingBranchFilter.cs
@@ -1,9 +1,9 @@
 using NuKeeper.Abstractions.CollaborationPlatform;
-using NuKeeper.Abstractions.DTOs;
 using NuKeeper.Abstractions.Logging;
 using NuKeeper.Inspection.RepositoryInspection;
 using System;
 using System.Threading.Tasks;
+using NuKeeper.Abstractions.CollaborationModels;
 
 namespace NuKeeper.Engine.Packages
 {
